requirements.txt
# Python server dependencies

# FastAPI web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Gradio (for Hugging Face Spaces)
gradio==4.16.0

# Image processing
Pillow==10.2.0
numpy==1.24.3

# BiRefNet model packages
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
accelerate==0.25.0
timm==0.9.12
einops==0.7.0
kornia==0.7.0

# ONNX conversion (optional)
# onnx==1.15.0
# onnxruntime==1.16.3